requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.29.0
flask-cors==3.0.10
